backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
crewai==0.28.0
langchain==0.1.6
langchain-community==0.0.20
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
tavily-python==0.3.0
requests==2.31.0
